process: add function to find process ids by listening port
