+++ measurement/util.py
@@ -129,35 +129,3 @@ def get_hostname():
     # This function is merely used, so only import socket package when necessary
     import socket
     return socket.gethostname()
-
-
-def find_process_by_port(port=None):
-    process_list = []
-    if not port:
-        logging.fatal("No process listening port specified.")
-        return
-
-    # literate over all file descriptors and build a socket address -> pid map
-    def build_inode_to_pid_map():
-        result = {}
-        for entry in os.scandir("/proc"):
-            # find all PIDs
-            if str.isdigit(entry.name):
-                try:
-                    for _fd in os.scandir("/proc/%s/fd" % entry.name):
-                        _fd_target = os.readlink(_fd.path)
-                        if not str.startswith(_fd_target, "socket"):
-                            continue
-                        _socket = _fd_target.split(":[")[:-2]
-                        try:
-                            result[_socket].append(entry.name)
-                        except KeyError:
-                            result[_socket] = [entry.name]
-                except PermissionError:
-                    logging.warn("Permission Denied reading /proc/%s/fd" % entry.name)
-        return result
-
-    def build_port_to_inode_map(port):
-        result = {}
-        
-        return
